+++ docs/components/PageHeader/PageHeader.tsx
@@ -1,20 +1,22 @@
-import { Badge, Container, Text, Title } from '@mantine/core';
-import { GithubIcon, NpmIcon } from '@mantinex/dev-icons';
+import { Badge, Container, Text, Title } from "@mantine/core";
+import { GithubIcon, NpmIcon } from "@mantinex/dev-icons";
 import {
   IconEdit,
   IconLicense,
   IconPackage,
   IconUserCode,
   IconVersions,
-} from '@tabler/icons-react';
-import pack from '../../../package/package.json';
-import type { PackageData } from '../../data';
-import classes from './PageHeader.module.css';
-import { PageHeaderLink } from './PageHeaderLink/PageHeaderLink';
+} from "@tabler/icons-react";
+import pack from "../../../package/package.json";
+import type { PackageData } from "../../data";
+import classes from "./PageHeader.module.css";
+import { PageHeaderLink } from "./PageHeaderLink/PageHeaderLink";
 
 interface PageHeaderProps {
   data: PackageData;
 }
+
+// https://github.com/gfazioli/mantine-flip/releases/tag/1.0.23
 
 export function PageHeader({ data }: PageHeaderProps) {
   return (
@@ -29,6 +31,13 @@ export function PageHeader({ data }: PageHeaderProps) {
             icon={<IconVersions size={18} stroke={1.5} />}
           >
             <Badge>v{pack.version}</Badge>
+          </PageHeaderLink>
+          <PageHeaderLink
+            label="Changelog"
+            icon={<GithubIcon size={16} />}
+            link={data.repositoryUrl + "/releases/tag/" + pack.version}
+          >
+            View the Changelog
           </PageHeaderLink>
           <PageHeaderLink
             label="Source"
@@ -63,7 +72,7 @@ export function PageHeader({ data }: PageHeaderProps) {
             icon={<IconUserCode size={18} stroke={1.5} />}
             link={`https://github.com/${data.author.githubUsername}`}
           >
-            {data.author.name}{' '}
+            {data.author.name}{" "}
             <Text span c="dimmed" inherit>
               (@{data.author.githubUsername})
             </Text>