+++ docs/demos/index.ts
@@ -1,6 +1,7 @@
 export { background } from "./Parallax.demo.background";
 export { configurator } from "./Parallax.demo.configurator";
 export { full } from "./Parallax.demo.full";
+export { initial } from "./Parallax.demo.initial";
 export { parallax } from "./Parallax.demo.parallax";
 export { shadow } from "./Parallax.demo.shadow";
 export { usage } from "./Parallax.demo.usage";