+++ CHANGELOG.md
@@ -3,6 +3,25 @@
 The following is a list of notable changes to the Mantine Parallax component.  
 Minor versions that are not listed in the changelog are bug fixes and small improvements.
 
+## 2.0.0 (2024-12-05)
+
+### ✨ New Features
+
+- ✨ Added `initialSkewX` and `initialSkewY` props
+- ✨ Added `initialRotationX`, `initialRotationY`, and `initialRotationZ` props
+- ✨ Added `initialPerspective` prop
+- ✨ Added `disabled` prop to disable the hover effect
+- ✨ Added `none` in the `perspective` and `initialPerspective` props when the value is greater or equal to `10000`
+
+### 📦 Other Changes
+
+- Updated the docs and demos with the new props
+- Updated the Mantine package to the latest version
+
+### 🐛 Bug Fixes
+- Minor improvements and bug fixes
+
+
 ## 1.0.0 (2024-11-29)
 
 - Initial release